+++ share/poly_test.go
@@ -11,7 +11,7 @@ import (
 
 func TestSecretRecovery(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 	poly := NewPriPoly(g, t, nil, g.RandomStream())
 	shares := poly.Shares(n)
@@ -29,22 +29,23 @@ func TestSecretRecovery(test *testing.T) {
 // tests the recovery of a secret when one of the share has an index
 // higher than the given `n`. This is a valid scenario that can happen during
 // a DKG-resharing:
-// 1. we add a new node n6 to an already-established group of 5 nodes.
-// 2. DKG runs without the first node in the group, i.e. without n1
-// 3. The list of qualified shares are [n2 ... n6] so the new resulting group
-//    has 5 members (no need to keep the 1st node around).
-// 4. When n6 wants to reconstruct, it will give its index given during the
+//  1. we add a new node n6 to an already-established group of 5 nodes.
+//  2. DKG runs without the first node in the group, i.e. without n1
+//  3. The list of qualified shares are [n2 ... n6] so the new resulting group
+//     has 5 members (no need to keep the 1st node around).
+//  4. When n6 wants to reconstruct, it will give its index given during the
+//
 // resharing, i.e. 6 (or 5 in 0-based indexing) whereas n = 5.
 // See TestPublicRecoveryOutIndex for testing with the commitment.
 func TestSecretRecoveryOutIndex(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 	poly := NewPriPoly(g, t, nil, g.RandomStream())
 	shares := poly.Shares(n)
 
 	selected := shares[n-t:]
-	require.Len(test, selected, t)
+	require.Len(test, selected, int(t))
 	newN := t + 1
 
 	recovered, err := RecoverSecret(g, selected, t, newN)
@@ -59,7 +60,7 @@ func TestSecretRecoveryOutIndex(test *testing.T) {
 
 func TestSecretRecoveryDelete(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 	poly := NewPriPoly(g, t, nil, g.RandomStream())
 	shares := poly.Shares(n)
@@ -82,7 +83,7 @@ func TestSecretRecoveryDelete(test *testing.T) {
 
 func TestSecretRecoveryDeleteFail(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 
 	poly := NewPriPoly(g, t, nil, g.RandomStream())
@@ -103,7 +104,7 @@ func TestSecretRecoveryDeleteFail(test *testing.T) {
 
 func TestSecretPolyEqual(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 
 	p1 := NewPriPoly(g, t, nil, g.RandomStream())
@@ -123,7 +124,7 @@ func TestSecretPolyEqual(test *testing.T) {
 
 func TestPublicCheck(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 
 	priPoly := NewPriPoly(g, t, nil, g.RandomStream())
@@ -139,7 +140,7 @@ func TestPublicCheck(test *testing.T) {
 
 func TestPublicRecovery(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 
 	priPoly := NewPriPoly(g, t, nil, g.RandomStream())
@@ -165,7 +166,7 @@ func TestPublicRecovery(test *testing.T) {
 
 func TestPublicRecoveryOutIndex(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 
 	priPoly := NewPriPoly(g, t, nil, g.RandomStream())
@@ -173,7 +174,7 @@ func TestPublicRecoveryOutIndex(test *testing.T) {
 	pubShares := pubPoly.Shares(n)
 
 	selected := pubShares[n-t:]
-	require.Len(test, selected, t)
+	require.Len(test, selected, int(t))
 	newN := t + 1
 
 	recovered, err := RecoverCommit(g, selected, t, newN)
@@ -195,7 +196,7 @@ func TestPublicRecoveryOutIndex(test *testing.T) {
 
 func TestPublicRecoveryDelete(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 
 	priPoly := NewPriPoly(g, t, nil, g.RandomStream())
@@ -220,7 +221,7 @@ func TestPublicRecoveryDelete(test *testing.T) {
 
 func TestPublicRecoveryDeleteFail(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 
 	priPoly := NewPriPoly(g, t, nil, g.RandomStream())
@@ -242,7 +243,7 @@ func TestPublicRecoveryDeleteFail(test *testing.T) {
 
 func TestPrivateAdd(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 
 	p := NewPriPoly(g, t, nil, g.RandomStream())
@@ -264,7 +265,7 @@ func TestPrivateAdd(test *testing.T) {
 
 func TestPublicAdd(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 
 	G := g.Point().Pick(g.RandomStream())
@@ -298,7 +299,7 @@ func TestPublicAdd(test *testing.T) {
 
 func TestPublicPolyEqual(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 
 	G := g.Point().Pick(g.RandomStream())
@@ -324,7 +325,7 @@ func TestPublicPolyEqual(test *testing.T) {
 
 func TestPriPolyMul(test *testing.T) {
 	suite := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 	a := NewPriPoly(suite, t, nil, suite.RandomStream())
 	b := NewPriPoly(suite, t, nil, suite.RandomStream())
@@ -351,7 +352,7 @@ func TestPriPolyMul(test *testing.T) {
 
 func TestRecoverPriPoly(test *testing.T) {
 	suite := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 	a := NewPriPoly(suite, t, nil, suite.RandomStream())
 
@@ -367,7 +368,7 @@ func TestRecoverPriPoly(test *testing.T) {
 	reverseRecovered, err := RecoverPriPoly(suite, reverses, t, n)
 	assert.Nil(test, err)
 
-	for i := 0; i < t; i++ {
+	for i := uint32(0); i < t; i++ {
 		assert.Equal(test, recovered.Eval(i).V.String(), a.Eval(i).V.String())
 		assert.Equal(test, reverseRecovered.Eval(i).V.String(), a.Eval(i).V.String())
 	}
@@ -375,12 +376,12 @@ func TestRecoverPriPoly(test *testing.T) {
 
 func TestPriPolyCoefficients(test *testing.T) {
 	suite := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 	a := NewPriPoly(suite, t, nil, suite.RandomStream())
 
 	coeffs := a.Coefficients()
-	require.Len(test, coeffs, t)
+	require.Len(test, coeffs, int(t))
 
 	b := CoefficientsToPriPoly(suite, coeffs)
 	require.Equal(test, a.coeffs, b.coeffs)
@@ -389,7 +390,7 @@ func TestPriPolyCoefficients(test *testing.T) {
 
 func TestRefreshDKG(test *testing.T) {
 	g := edwards25519.NewBlakeSHA256Ed25519()
-	n := 10
+	n := uint32(10)
 	t := n/2 + 1
 
 	// Run an n-fold Pedersen VSS (= DKG)
@@ -397,7 +398,7 @@ func TestRefreshDKG(test *testing.T) {
 	priShares := make([][]*PriShare, n)
 	pubPolys := make([]*PubPoly, n)
 	pubShares := make([][]*PubShare, n)
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		priPolys[i] = NewPriPoly(g, t, nil, g.RandomStream())
 		priShares[i] = priPolys[i].Shares(n)
 		pubPolys[i] = priPolys[i].Commit(nil)
@@ -405,8 +406,8 @@ func TestRefreshDKG(test *testing.T) {
 	}
 
 	// Verify VSS shares
-	for i := 0; i < n; i++ {
-		for j := 0; j < n; j++ {
+	for i := uint32(0); i < n; i++ {
+		for j := uint32(0); j < n; j++ {
 			sij := priShares[i][j]
 			// s_ij * G
 			sijG := g.Point().Base().Mul(sij.V, nil)
@@ -416,9 +417,9 @@ func TestRefreshDKG(test *testing.T) {
 
 	// Create private DKG shares
 	dkgShares := make([]*PriShare, n)
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		acc := g.Scalar().Zero()
-		for j := 0; j < n; j++ { // assuming all participants are in the qualified set
+		for j := uint32(0); j < n; j++ { // assuming all participants are in the qualified set
 			acc = g.Scalar().Add(acc, priShares[j][i].V)
 		}
 		dkgShares[i] = &PriShare{i, acc}
@@ -426,9 +427,9 @@ func TestRefreshDKG(test *testing.T) {
 
 	// Create public DKG commitments (= verification vector)
 	dkgCommits := make([]kyber.Point, t)
-	for k := 0; k < t; k++ {
+	for k := uint32(0); k < t; k++ {
 		acc := g.Point().Null()
-		for i := 0; i < n; i++ { // assuming all participants are in the qualified set
+		for i := uint32(0); i < n; i++ { // assuming all participants are in the qualified set
 			_, coeff := pubPolys[i].Info()
 			acc = g.Point().Add(acc, coeff[k])
 		}
@@ -437,7 +438,7 @@ func TestRefreshDKG(test *testing.T) {
 
 	// Check that the private DKG shares verify against the public DKG commits
 	dkgPubPoly := NewPubPoly(g, nil, dkgCommits)
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		require.True(test, dkgPubPoly.Check(dkgShares[i]))
 	}
 
@@ -448,7 +449,7 @@ func TestRefreshDKG(test *testing.T) {
 	subPubShares := make([][]*PubShare, n)
 
 	// Create subshares and subpolys
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		subPriPolys[i] = NewPriPoly(g, t, dkgShares[i].V, g.RandomStream())
 		subPriShares[i] = subPriPolys[i].Shares(n)
 		subPubPolys[i] = subPriPolys[i].Commit(nil)
@@ -458,10 +459,10 @@ func TestRefreshDKG(test *testing.T) {
 
 	// Handout shares to new nodes column-wise and verify them
 	newDKGShares := make([]*PriShare, n)
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		tmpPriShares := make([]*PriShare, n) // column-wise reshuffled sub-shares
 		tmpPubShares := make([]*PubShare, n) // public commitments to old DKG private shares
-		for j := 0; j < n; j++ {
+		for j := uint32(0); j < n; j++ {
 			// Check 1: Verify that the received individual private subshares s_ji
 			// is correct by evaluating the public commitment vector
 			tmpPriShares[j] = &PriShare{I: j, V: subPriShares[j][i].V} // Shares that participant i gets from j
@@ -486,9 +487,9 @@ func TestRefreshDKG(test *testing.T) {
 
 	// Refresh the DKG commitments (= verification vector)
 	newDKGCommits := make([]kyber.Point, t)
-	for i := 0; i < t; i++ {
+	for i := uint32(0); i < t; i++ {
 		pubShares := make([]*PubShare, n)
-		for j := 0; j < n; j++ {
+		for j := uint32(0); j < n; j++ {
 			_, c := subPubPolys[j].Info()
 			pubShares[j] = &PubShare{I: j, V: c[i]}
 		}
@@ -501,13 +502,13 @@ func TestRefreshDKG(test *testing.T) {
 	require.True(test, dkgCommits[0].Equal(newDKGCommits[0]))
 
 	// Check that the old and new DKG private shares are different
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		require.False(test, dkgShares[i].V.Equal(newDKGShares[i].V))
 	}
 
 	// Check that the refreshed private DKG shares verify against the refreshed public DKG commits
 	q := NewPubPoly(g, nil, newDKGCommits)
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		require.True(test, q.Check(newDKGShares[i]))
 	}
